requirements.txt
azure-functions
azure-storage-blob==12.23.0